remove warning

+++ Game/GameCore.cpp
@@ -164,9 +164,9 @@ void	GameCore::bombManager(Actions &act)
 					}
 			}
 			_updateEntities.push_back(std::unique_ptr<IEntity>(&a));
-			if (a.getOwner() == _player1.getId())
+			if (a.getOwner() == static_cast<unsigned int>(_player1.getId()))
 				_player1.addBomb();
-			if (a.getOwner() == _player2.getId())
+			if (a.getOwner() == static_cast<unsigned int>(_player2.getId()))
 				_player2.addBomb();
 			break;
 		}
